pyproject.toml
[tool.poetry]
name = "refidxdb"
version = "0.0.1"
description = "Python interface for various refractive index databases"
authors = ["Mirza Arnaut <mirza.arnaut45@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
polars = "^1.13.1"
numpy = "^2.1.3"
click = "^8.1.7"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
refidxdb = "refidxdb.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
